renamed PromoHeader folder
